Refactor: added new local provider: LM Studio. Still testing, not functioning as expected as yet.
